Some Table Columns Update

+++ Hotel-backend/Common/Dto/MarketingDecisionDto.cs
@@ -13,8 +13,8 @@ public class MarketingDecisionDto
     public int GroupID { get; set; }
     public string MarketingTechniques { get; set; }
     public string Segment { get; set; }
-    public int Spending { get; set; }
-    public int LaborSpending { get; set; }
+    public decimal Spending { get; set; }
+    public decimal LaborSpending { get; set; }
     public int ActualDemand { get; set; }
     public bool Confirmed { get; set; }
